Rewrite Sidebar in storages to SidePanel component
