+++ modules/storages/app/components/storages/admin/side_panel/health_notifications_component.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 #-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) 2012-2024 the OpenProject GmbH
@@ -30,29 +28,29 @@
 
 module Storages
   module Admin
-    module Sidebar
-      class ValidationResultComponent < ApplicationComponent # rubocop:disable OpenProject/AddPreviewForViewComponent
-        include OpPrimer::ComponentHelpers
-        include OpTurbo::Streamable
+    class SidePanel::HealthNotificationsComponent < ApplicationComponent
+      include ApplicationHelper
+      include OpTurbo::Streamable
+      include OpPrimer::ComponentHelpers
 
-        def initialize(result:)
-          super(result)
-          @result = result
-        end
+      def initialize(storage:)
+        super
+        @storage = storage
+      end
 
-        private
+      def render?
+        @storage.automatically_managed?
+      end
 
-        def status_indicator
-          case @result.type
-          when :healthy
-            { scheme: :success, label: I18n.t("storages.health.label_healthy") }
-          when :warning
-            { scheme: :attention, label: I18n.t("storages.health.label_warning") }
-          when :error
-            { scheme: :danger, label: I18n.t("storages.health.label_error") }
-          else
-            raise ArgumentError, "Unknown validation result type for status indicator: #{@result.type}"
-          end
+      def notification_status
+        if @storage.health_notifications_should_be_sent?
+          { icon: :"bell-slash",
+            label: I18n.t("storages.health_email_notifications.unsubscribe"),
+            description: I18n.t("storages.health_email_notifications.description_subscribed") }
+        else
+          { icon: :bell,
+            label: I18n.t("storages.health_email_notifications.subscribe"),
+            description: I18n.t("storages.health_email_notifications.description_unsubscribed") }
         end
       end
     end
