+++ modules/storages/app/components/storages/admin/side_panel/validation_result_component.html.erb
@@ -29,30 +29,28 @@ See COPYRIGHT and LICENSE files for more details.
 
 <%=
   component_wrapper do
-    flex_layout do |container|
-      container.with_row do
-        render(Primer::Beta::Text.new(font_weight: :bold, test_selector: "validation-result--subtitle")) do
-          I18n.t("storages.health.connection_validation.subtitle")
+    render(Primer::OpenProject::SidePanel::Section.new) do |section|
+      section.with_title(test_selector: "validation-result--subtitle") { I18n.t("storages.health.connection_validation.subtitle") }
+
+      flex_layout do |container|
+        if @result.validation_result_exists?
+          container.with_row(mt: 2) do
+            status = status_indicator
+
+            concat(render(Primer::Beta::Text.new(pr: 2, test_selector: "validation-result--timestamp")) do
+              I18n.t('storages.health.checked', datetime: helpers.format_time(@result.timestamp))
+            end)
+            concat(render(Primer::Beta::Label.new(scheme: status[:scheme])) { status[:label] })
+          end
         end
-      end
 
-      if @result.validation_result_exists?
-        container.with_row(mt: 2) do
-          status = status_indicator
+        if @result.description.present?
+          prefix = @result.error_code? ? "#{@result.error_code.upcase}: " : ""
 
-          concat(render(Primer::Beta::Text.new(pr: 2, test_selector: "validation-result--timestamp")) do
-            I18n.t('storages.health.checked', datetime: helpers.format_time(@result.timestamp))
-          end)
-          concat(render(Primer::Beta::Label.new(scheme: status[:scheme])) { status[:label] })
-        end
-      end
-
-      if @result.description.present?
-        prefix = @result.error_code? ? "#{@result.error_code.upcase}: " : ""
-
-        container.with_row(mt: 2) do
-          render(Primer::Beta::Text.new(color: :muted, test_selector: "validation-result--description")) do
-            prefix + @result.description
+          container.with_row(mt: 2) do
+            render(Primer::Beta::Text.new(color: :muted, test_selector: "validation-result--description")) do
+              prefix + @result.description
+            end
           end
         end
       end
