+++ modules/storages/app/components/storages/admin/side_panel_component.html.erb
@@ -1,0 +1,11 @@
+<%=
+  component_wrapper do
+    render(Primer::OpenProject::SidePanel.new) do |panel|
+      panel.with_section(Storages::Admin::SidePanel::HealthStatusComponent.new(storage: @storage))
+
+      if @storage.automatic_management_enabled?
+        panel.with_section(Storages::Admin::SidePanel::HealthNotificationsComponent.new(storage: @storage))
+      end
+    end
+  end
+%>
