--- modules/storages/app/components/storages/admin/sidebar_component.html.erb
@@ -1,11 +1,0 @@
-<%=
-  component_wrapper do
-    render(Primer::OpenProject::BorderGrid.new) do |border_grid|
-      border_grid.with_row { render(Storages::Admin::Sidebar::HealthStatusComponent.new(storage: @storage)) }
-
-      if @storage.automatic_management_enabled?
-        border_grid.with_row { render(Storages::Admin::Sidebar::HealthNotificationsComponent.new(storage: @storage)) }
-      end
-    end
-  end
-%>
