+++ modules/storages/app/controllers/storages/admin/storages_controller.rb
@@ -188,7 +188,7 @@ class Storages::Admin::StoragesController < ApplicationController
     return head :bad_request unless %w[1 0].include?(permitted_storage_params[:health_notifications_enabled])
 
     if @storage.update(health_notifications_enabled: permitted_storage_params[:health_notifications_enabled])
-      update_via_turbo_stream(component: Storages::Admin::Sidebar::HealthNotificationsComponent.new(storage: @storage))
+      update_via_turbo_stream(component: Storages::Admin::SidePanel::HealthNotificationsComponent.new(storage: @storage))
       respond_with_turbo_streams
     else
       flash.now[:primer_banner] = {
